+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-beta.xx]-2025-03-01
+
+### Added
+
+- Orders API endpoints:
+  - Download Invoice
+  - Get order by id
+  - Get order by order number
+  - Get list of orders
+  - Get list of booking orders
+
 ## [1.0.0-beta.7]-2025-02-28
 
 ### Added